+++ overwatch/processing/detectors/TPC.py
@@ -22,27 +22,78 @@ from .. import processingClasses
 # Trending Classes
 ##################
 class TPCTrendingObjectMean(processingClasses.trendingObject):
+    """ Basic trending object which extracts the mean from a 1D histogram.
+
+    This function serves as a proof of principle for a trending object which trends the mean
+    value of a 1D histogram vs time. Although it is labeled as TPC, it is actually entirely
+    general. It is just labeled as such here because that was its original purpose.
+
+    Most of the heavy lifting is done by the base class. This class simply implements the methods
+    to extract the values (values are extracted via the dedicated histogram methods, so this class
+    is rather light).
+
+    Args:
+        trendingHistName (str): Name of the trending object.
+        trendingHistTitle (str): Name of the trending object that is appropriate for display.
+        histNames (list): List of the names of histograms which are needed to perform the trending.
+        nEntries (int): Number of entries the trending object should contain. Default: 50.
+
+    Attributes:
+        Attributes of ``trendingObject``. No attributes are added by this derived class.
+    """
     def __init__(self, trendingHistName, trendingHistTitle, histNames, nEntries = 50):
         super(TPCTrendingObjectMean, self).__init__(trendingName = trendingHistName, prettyTrendingName = trendingHistTitle, nEntries = nEntries, trendingHist = None, histNames = histNames)
 
-    def retrieveHist(self):
-        super(TPCTrendingObjectMean, self).retrieveHist()
+    def retrieveHistogram(self):
+        """ Retrieve or create a graph based on the stored numpy array.
 
-        # Set the histogrma to display a time axis
+        This function relies on the base class to perform the actual retrieval. This function
+        provides additional customization to the object. In particular, we set the x-axis to
+        display a time axis, as appropriate for a trended object. We could also customize the
+        format of the time stored on the x-axis.
+
+        Args:
+            None
+        Returns:
+            histogramContainer: Container which holds the created graph. It is returned to allow for
+                further customization. This histogram container is already stored in the object.
+        """
+        # This will retrieve the underlying object from the database. Afterwards retrieval,
+        # we can customized it.
+        super(TPCTrendingObjectMean, self).retrieveHistogram()
+
+        # Set the histogram to display a time axis
         self.hist.hist.GetXaxis().SetTimeDisplay(1)
+        # Could also configure the time format.
         #self.trendingHist.GetXaxis().SetTimeFormat()
 
-        # Make it more visible
+        # Make the markers more visible
         self.hist.hist.SetMarkerStyle(ROOT.kFullCircle)
 
-        print("self.hist: {}, self.hist.hist: {}".format(self.hist, self.hist.hist))
+        logger.debug("self.hist: {}, self.hist.hist: {}".format(self.hist, self.hist.hist))
+
+        # The hist is already available through the histogram container, but we return the hist
+        # container in case the caller wants to do additional customization
+        return self.hist
 
     def fill(self, hist):
-        if len(self.histNames) > 1:
-            print("Too many histograms passed to {0}!".format(self.histNames))
-            return
+        """ Extract the mean and error and store the value in the trending objects.
 
-        print("Filling hist {}, histNames: {}".format(self.hist, self.histNames))
+        The values are extracted directly from the histograms via their dedicated methods. Storage of the
+        values is passed onto the ``trendingObject.fill(...)`` method of the base class. Note that the
+        method signatures are different!
+
+        The name of this method was inspired by ``TH1.Fill(val)`` in which it fills values into a histogram.
+
+        Args:
+            hist (histogramContainer): Histogram from which the trended value should be extracted.
+        Returns:
+            None. The extracted value is stored in the trending object.
+        """
+        if len(self.histNames) > 1:
+            raise ValueError("Too many histograms passed to the trending object. Expected histograms named: {}!".format(self.histNames))
+
+        logger.debug("Filling hist {}, histNames: {}".format(self.hist, self.histNames))
         fillVal = 0
         fillValError = 0
         #for histName in self.histNames:
@@ -51,11 +102,34 @@ class TPCTrendingObjectMean(processingClasses.trendingObject):
         fillVal += hist.hist.GetMean()
         fillValError += hist.hist.GetMeanError()
 
-        print("Filling value: {}, error: {}".format(fillVal, fillValError))
+        logger.debug("Filling value: {}, error: {}".format(fillVal, fillValError))
         super(TPCTrendingObjectMean, self).fill(fillVal, fillValError)
 
 def defineTPCTrendingObjects(trending, **kwargs):
+    """ Define the TPC trending objects, including the histogram upon which they rely (via the histogram names).
+
+    In particular, we need to define the trending objects, with their names as the keys in the trending
+    dictionary. Of particular note are the histogram names which are needed to contribute to the trending
+    object. Since in principle this could be more than one histogram, the names must be specified in a list.
+    Since the TPC trending histograms only need one histogram, the list is just one entry long.
+
+    While we could define the objects one by one, it would lead to some repeated code. Instead, we define
+    a set of nested lists for convenience. The format of the each entry in the list is
+    ``["trendingObjectName", "Trending Object Display Name", ["histogramName"]]``. This way, we only call
+    the definition of the trending object once, which allows us to easily change it if necessary. We are able
+    to utilize this approach because every object defined here is the trend of the mean of a particular
+    quantity.
+
+    Args:
+        trending (dict): Where the trending objects should be defined. Keys are the name of the trending
+            objects, while values are the trending objects themselves.
+        **kwargs (dict): Reserved for future use.
+    Returns:
+        dict: Dictionary filled with the trending objects. Keys are the name of the trending objects, while
+            values are the trending objects themselves.
+    """
     # Being a bit clever so we don't have to repeat too much code
+    # Each list entry is ["trendingObjectName", "Trending Object Display Name", ["histogramName"]]
     names = [["TPCClusterTrending", "<TPC clusters>: (p_{T} > 0.25 GeV/c, |#eta| < 1)", ["TPCQA/h_tpc_track_all_recvertex_0_5_7_restrictedPtEta"]],
              ["TPCFoundClusters", "<Found/Findable TPC clusters>: (p_{T} > 0.25 GeV/c, |#eta| < 1)", ["TPCQA/h_tpc_track_all_recvertex_2_5_7_restrictedPtEta"]],
              ["TPCdcaR", "<DCAr> (cm)>: (p_{T}> 0.25 GeV/c, |#eta| < 1)", ["TPCQA/h_tpc_track_all_recvertex_3_5_7_restrictedPtEta"]],
@@ -66,10 +140,12 @@ def defineTPCTrendingObjects(trending, **kwargs):
              ["histMpos", "<Multiplicity of pos. tracks>", ["TPCQA/h_tpc_event_recvertex_4"]],
              ["histMneg", "<Multiplicity of neg. tracks>", ["TPCQA/h_tpc_event_recvertex_5"]]]
 
+    # Create and store the object
     for name, title, histNames in names:
-        # Create it if it doens't exist
+        # It may be possible that the object already exists, so be certain that we don't overwrite it.
+        # NOTE: The object already existing is not possible as of August 2018.
         if not name in trending.keys():
-            # Define new trending histogram
+            # Define and store new trending histogram
             trending[name] = TPCTrendingObjectMean(name, title, histNames)
 
     return trending